+++ tests/SurfaceMeshTest.cpp
@@ -348,18 +348,6 @@ TEST_F(SurfaceMeshTest, is_manifold)
         EXPECT_TRUE(mesh.is_manifold(v));
 }
 
-TEST_F(SurfaceMeshTest, edge_length)
-{
-    add_quad();
-    Scalar sum(0);
-    for (auto e : mesh.edges())
-    {
-        sum += mesh.edge_length(e);
-    }
-    sum /= (Scalar)mesh.n_edges();
-    EXPECT_FLOAT_EQ(sum, 1.0);
-}
-
 TEST_F(SurfaceMeshTest, property_stats)
 {
     mesh.property_stats();